refactor: split engine/mod.rs into focused sub-modules
Break the monolithic src/engine/mod.rs into eight modules:

- rule.rs: SigmaRule, Detection, Selection, FieldCriterion, serde structs
- logsource.rs: LogSource, LogSourceKey, platform support, classification
- condition.rs: condition parsing, transpilation, compiled condition tree
- matcher.rs: pattern matching, field matching, numeric/CIDR/regex ops
- loader.rs: filesystem traversal, YAML parsing, rule compile pipeline
- alert.rs: Sigma alert and match-detail construction
- stats.rs: EngineStats and rule counters
- handler.rs: SigmaDetectionHandler (via mod.rs re-export)

Public imports through crate::engine remain unchanged.
Sigma semantics and alert output are unaffected.
